normalement c'est ok !

+++ src/servlets/DeleteContactServlet.java
@@ -11,6 +11,7 @@ import javax.servlet.http.HttpServletResponse;
 import org.springframework.context.ApplicationContext;
 import org.springframework.web.context.support.WebApplicationContextUtils;
 
+import domain.metier.Contact;
 import domain.services.interfaces.IContactService;
 
 /**
@@ -30,6 +31,7 @@ public class DeleteContactServlet extends HttpServlet {
 		System.out.println("DeleteContact doPost");
 		
 		String id = request.getParameter("id");
+		String version = request.getParameter("version");
 		
 		/*TODO: vérification confirmité des champs*/
 		boolean okId = id!=null && id.length()>0;
@@ -43,12 +45,21 @@ public class DeleteContactServlet extends HttpServlet {
 		if(okId){
 			ApplicationContext context = WebApplicationContextUtils.getWebApplicationContext(getServletContext());
 			IContactService contactService = (IContactService) context.getBean("contactService");
-			contactService.deleteContact(idLong);
-			request.setAttribute("message", "Contact with id n°"+id+" has been correctly delete !");
-			RequestDispatcher dispatcher = request.getRequestDispatcher("MyContactServlet");
-			dispatcher.forward(request, response);
+			Contact c = contactService.getContactById(Long.parseLong(id));
+			
+			if(c==null || version!=null && c.getVersion()!=Long.parseLong(version)){
+				request.setAttribute("success", false);
+				request.setAttribute("message", "Contact not deleted (not up-to-date).");
+			} else {
+				contactService.deleteContact(idLong);
+				request.setAttribute("message", "Contact with id n°"+id+" has been delete !");
+				request.setAttribute("success", true);
+				RequestDispatcher dispatcher = request.getRequestDispatcher("MyContactServlet");
+				dispatcher.forward(request, response);
+			}
 		} else {
 			request.setAttribute("message", "Error with the field...");
+			request.setAttribute("success", false);
 			RequestDispatcher dispatcher = request.getRequestDispatcher("MyContactServlet");
 			dispatcher.forward(request, response);
 		}
